% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{beta_plots}
\alias{beta_plots}
\title{Plots of coefficients by lambda for cv.rq.group.pen and cv.rq.pen}
\usage{
beta_plots(model, voi = NULL, logLambda = TRUE, loi = NULL, ...)
}
\arguments{
\item{model}{cv.rq.pen or cv.rq.group.pen object}

\item{voi}{Index of betas to include. Default is all of them.}

\item{logLambda}{Plot of lambdas is on the log scale.}

\item{loi}{Index of lambdas to use, default is all of them.}

\item{...}{Additional arguments to be sent to plot()}
}
\value{
Plot of how beta estimates change with lambda.
}
\description{
Warning: this function is deprecated and will not be exported in future versions.
}
\examples{
  set.seed(1)
  x <- matrix(rnorm(800),nrow=100)
  y <- 1 + x[,1] - 3*x[,5] + rnorm(100)
  lassoModels <- cv.rq.pen(x,y)
  b_plot <- beta_plots(lassoModels)
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
