% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{bytau.plot}
\alias{bytau.plot}
\title{Plot of how coefficients change with tau}
\usage{
bytau.plot(x, ...)
}
\arguments{
\item{x}{A rq.pen.seq or rq.pen.seq.cv object.}

\item{...}{Additional arguments see bytau.plot.rq.pen.seq() or bytau.plot.rq.pen.seq.cv() for more information.}
}
\value{
Returns the plot of how coefficients change with tau.
}
\description{
Plot of how coefficients change with tau
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
