% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupQPenCv.R
\name{coef.cv.hrq_tau_glasso}
\alias{coef.cv.hrq_tau_glasso}
\title{Title}
\usage{
\method{coef}{cv.hrq_tau_glasso}(cv.fit, s)
}
\arguments{
\item{cv.fit}{cv object from \code{cv.hrq_tau_glasso}}

\item{s}{optimal lambda value}
}
\value{
coefficients
}
\description{
Title
}
