% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{coef.cv.rq.pen}
\alias{coef.cv.rq.pen}
\title{Returns Coefficients of a cv.rq.pen object}
\usage{
\method{coef}{cv.rq.pen}(object, lambda = "min", ...)
}
\arguments{
\item{object}{cv.rq.pen object}

\item{lambda}{Value of lambda, default is to use the minimum value.}

\item{...}{Additional parameters.}
}
\value{
Coefficients for a given lambda, or the lambda associated with the minimum cv value.
}
\description{
Warning: this function will be deprecated and not exported in future versions of rqPen, due to the switch from cv.rq.pen() to rq.pen.cv().
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
