% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{cv_plots}
\alias{cv_plots}
\title{Plots of cross validation results as a function of lambda.}
\usage{
cv_plots(model, logLambda = TRUE, loi = NULL, ...)
}
\arguments{
\item{model}{A cv.rq.pen() object.}

\item{logLambda}{Whether lambda values should be logged or not.}

\item{loi}{Lambda indexes of interest, if null all lambda values will be used.}

\item{...}{Additional parameters sent to plot function.}
}
\value{
returns a cross validation plot
}
\description{
Plots of cross validation results as a function of lambda.
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
\keyword{internal}
