% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{plot.cv.rq.group.pen}
\alias{plot.cv.rq.group.pen}
\title{Cross validation plot for cv.rq.group.pen object}
\usage{
\method{plot}{cv.rq.group.pen}(x, ...)
}
\arguments{
\item{x}{A cv.rq.group.pen object}

\item{...}{Additional parameters for plot function.}
}
\value{
A cross validation plot.
}
\description{
Cross validation plot for cv.rq.group.pen object
}
\keyword{internal}
