% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrq_tau_glasso_predict.R
\name{predict.cv.hrq_tau_glasso}
\alias{predict.cv.hrq_tau_glasso}
\title{Title Prediction (from CV object) for quantile regression and consistent variable selection}
\usage{
\method{predict}{cv.hrq_tau_glasso}(fit, newX = NULL, s = NULL)
}
\arguments{
\item{fit}{model object from \code{hrq_tau_glasso}}

\item{newX}{new covariate matrix. If not supplied, fitted value is returned.}

\item{s}{optimal value of lambda. If not supplied, a list of prediction is returned, where each element of the list corresponds to a lambda.}
}
\value{
predicted response for each tau
}
\description{
Title Prediction (from CV object) for quantile regression and consistent variable selection
}
