% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{predict.cv.rq.pen}
\alias{predict.cv.rq.pen}
\title{Prediction for a cv.rq.pen object}
\usage{
\method{predict}{cv.rq.pen}(object, newx, lambda = "lambda.min", ...)
}
\arguments{
\item{object}{A cv.rq.pen object.}

\item{newx}{Matrix of new data to make predictions with.}

\item{lambda}{Lambda value used, default is the value associated with the minimum cross validation result.}

\item{...}{Additional parameters that are currenlty ignored}
}
\value{
A vector of predictions.
}
\description{
This function is deprecated and will not be exported in future versions.
}
