% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupQPenPredict.R
\name{predict.hrq_tau_glasso}
\alias{predict.hrq_tau_glasso}
\title{Title Prediction for quantile regression and consistent variable selection}
\usage{
\method{predict}{hrq_tau_glasso}(fit, newX, s = NULL)
}
\arguments{
\item{fit}{model object from \code{hrq_tau_glasso}}

\item{newX}{new covariate matrix}

\item{s}{optimal value of lambda}
}
\value{
predicted response for each tau
}
\description{
Title Prediction for quantile regression and consistent variable selection
}
