% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{predict.qic.select}
\alias{predict.qic.select}
\title{Predictions from a qic.select object}
\usage{
\method{predict}{qic.select}(object, newdata, ...)
}
\arguments{
\item{object}{qic.select object}

\item{newdata}{Data matrix to make predictions from.}

\item{...}{optional arguments}
}
\value{
A matrix of predicted values.
}
\description{
Predictions from a qic.select object
}
\examples{
x <- matrix(runif(800),ncol=8)
y <- 1 + x[,1] + x[,8] + (1+.5*x[,3])*rnorm(100)
m1 <- rq.pen(x,y,tau=c(.25,.75))
q1 <- qic.select(m1)
newx <- matrix(runif(80),ncol)
preds <- predict(q1,newx)
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
