% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{print.cv.rq.pen}
\alias{print.cv.rq.pen}
\title{Prints a cv.rq.pen object.}
\usage{
print.cv.rq.pen(x, ...)
}
\arguments{
\item{x}{cv.rq.pen object}

\item{...}{Optional arguments, not used.}
}
\value{
Prints cross validation or information criterion values by lambda.
}
\description{
Prints a cv.rq.pen object.
}
\details{
Warning this function is no longer exported.
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
