% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{print.qic.select}
\alias{print.qic.select}
\title{Print a qic.select object}
\usage{
print.qic.select(x, ...)
}
\arguments{
\item{x}{qic.select object}

\item{...}{optional arguments}
}
\value{
Prints the coefficients of the qic.select object
}
\description{
Print a qic.select object
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
