% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{print.rq.pen}
\alias{print.rq.pen}
\title{Prints an rq.pen object}
\usage{
\method{print}{rq.pen}(x, ...)

\method{print}{rq.pen}(x, ...)
}
\arguments{
\item{x}{A rq.pen object}

\item{...}{Additional arguments}
}
\value{
Prints the coefficients of the object.


}
\description{
Warning this function is deprecated and will not be exported in future releases.
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
