% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{print.rq.pen.seq.cv}
\alias{print.rq.pen.seq.cv}
\title{Prints a rq.pen.seq.cv object}
\usage{
print.rq.pen.seq.cv(x, ...)
}
\arguments{
\item{x}{A req.pen.seq.cv object.}

\item{...}{Additional arguments.}
}
\value{
Print of btr and gtr from a rq.pen.seq.cv object. If only one quantile is modeled then only btr is returned.
}
\description{
Prints a rq.pen.seq.cv object
}
