% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{print.rq.pen.seq}
\alias{print.rq.pen.seq}
\title{Print a rq.pen.seq object}
\usage{
\method{print}{rq.pen.seq}(x, ...)
}
\arguments{
\item{x}{rq.pen.seq object}

\item{...}{optional arguments}
}
\value{
If only one model, prints a data.frame of the number of nonzero coefficients and lambda. Otherwise prints information about the quantiles being modeled and choices for a.
}
\description{
Print a rq.pen.seq object
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
