% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{qic.select}
\alias{qic.select}
\title{Select model using IC}
\usage{
qic.select(x, ...)
}
\arguments{
\item{x}{A rq.pen.seq or rq.pen.seq.cv object.}

\item{...}{Additional arguments see bytau.plot.rq.pen.seq() or bytau.plot.rq.pen.seq.cv() for more information.}
}
\value{
Returns the plot of how coefficients change with tau.
}
\description{
Select model using IC
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
