% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqPen.R
\docType{package}
\name{rqPen}
\alias{rqPen}
\alias{rqPen-package}
\title{rqPen: A package for estimating quantile regression models using penalized objective functions.}
\description{
The package estimates a quantile regression model using LASSO, Adaptive LASSO, SCAD, MCP, elastic net, 
and their group counterparts, with the exception of elastic net for which there is no group penalty implementation.
}
\section{rqPen functions}{

The most important functions are rq.pen(), rq.group.pen(), rq.pen.cv() and rq.group.pen.cv(). These functions 
fit quantile regression models with individual or group penalties. The cv functions automate the cross-validation process for selection of tuning parameters.
}

