% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someCode.R
\name{wNorm}
\alias{wNorm}
\title{Weighted Norm example}
\usage{
wNorm(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The product of x and y.
}
\description{
This function multiplies two numbers.
}
\examples{
wNorm(2, 3)

}
