% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrq_tau_glasso_cv.R
\name{coef.cv.hrq_tau_glasso}
\alias{coef.cv.hrq_tau_glasso}
\title{Title Getting the coefficient estimates from \code{cv.hrq_tau_glasso}.}
\usage{
\method{coef}{cv.hrq_tau_glasso}(cv.fit, s)
}
\arguments{
\item{cv.fit}{The CV object from \code{cv.hrq_tau_glasso}}

\item{s}{lambda value, or can be character either "lambda.min" or "lambda.1se". If not specified, "lambda.min" is used.}
}
\value{
coefficient estimates corresponding to the specified lambda values.
}
\description{
Title Getting the coefficient estimates from \code{cv.hrq_tau_glasso}.
}
