% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{coef.cv.rq.group.pen}
\alias{coef.cv.rq.group.pen}
\title{Coefficients from a cv.rq.group.pen object}
\usage{
\method{coef}{cv.rq.group.pen}(object, lambda = "min", ...)
}
\arguments{
\item{object}{A cv.rq.group.pen object.}

\item{lambda}{The lambda value, default is to use the one associated with the minimum cv error.}

\item{...}{Additional parameters.}
}
\value{
Vector of coefficients.
}
\description{
Coefficients from a cv.rq.group.pen object
}
