% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{cv_plots}
\alias{cv_plots}
\title{Plots of cross validation results as a function of lambda.}
\usage{
cv_plots(model, logLambda = TRUE, loi = NULL, ...)
}
\arguments{
\item{model}{}

\item{logLambda}{}

\item{loi}{}

\item{...}{}
}
\value{
returns a cross validation plot
}
\description{
Plots of cross validation results as a function of lambda.
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
