% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{predict.rq.pen}
\alias{predict.rq.pen}
\title{Prediction for a rq.pen object}
\usage{
\method{predict}{rq.pen}(object, newx, ...)
}
\arguments{
\item{object}{An rq.pen object.}

\item{newx}{Matrix of new data to make predictions with.}

\item{...}{Additional parameters that are currenlty ignored}
}
\value{
A vector of predictions.
}
\description{
This function is no longer exported.
}
\keyword{internal}
