\name{print.cv.rq.pen}
\alias{print.cv.rq.pen}
\title{Print cv.rq.pen object}
\usage{
\method{print}{cv.rq.pen}(x,...)
}
\arguments{
  \item{x}{Object to be printed.}
  \item{...}{Additional arguments for coefficient function}
}
\description{
Prints the cross validation (or bic) results and reports coefficients for a cv.rq.pen object. Default coefficients are for value that minimizes 
cross validation or BIC. 
}
\author{Ben Sherwood}