% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{print.cv.rq.pen}
\alias{print.cv.rq.pen}
\title{Prints a cv.rq.pen object.}
\usage{
\method{print}{cv.rq.pen}(x, ...)

\method{print}{cv.rq.pen}(x, ...)
}
\arguments{
\item{x}{A cv.rq.pen object}

\item{...}{Additional arguments}
}
\value{
Prints cross validation or information criterion values by lambda.

Prints coefficients and cross validation results.
}
\description{
Warning: this function is deprecated and will not be exported in future releases.
}
\details{
Warning this function is deprecated and will not be exported in future releases.
}
\author{
Ben Sherwood, \email{ben.sherwood@ku.edu}
}
