% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{qic.select}
\alias{qic.select}
\title{Title}
\usage{
qic.select(obj, method = "BIC", septau = FALSE, weights = NULL)
}
\arguments{
\item{obj}{A rq.pen.seq or rq.pen.seq.cv object.}

\item{method}{Choice of BIC, AIC or PBIC, a large p BIC.}

\item{septau}{If optimal}

\item{weights}{}
}
\value{

}
\description{
Title
}
