% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gqCvPen.R
\name{rq.gq.pen.cv}
\alias{rq.gq.pen.cv}
\title{Title Cross validation for consistent variable selection across multiple quantiles}
\usage{
rq.gq.pen.cv(
  x = NULL,
  y = NULL,
  tau = NULL,
  lambda = NULL,
  nfolds = 10,
  loss = c("rq", "se"),
  wt_tau_loss = NULL,
  foldid = NULL,
  ...
)
}
\arguments{
\item{x}{covariate matrix. Not needed if \code{model_obj} is supplied.}

\item{y}{univariate response. Not needed if \code{model_obj} is supplied.}

\item{tau}{a sequence of tau to be modeled}

\item{lambda}{The sequence of lambdas.}

\item{nfolds}{number of folds}

\item{loss}{loss function to be evaluated. Supported loss functions include quantile ("rq") and squared loss("se"). Default is the quantile loss.}

\item{wt_tau_loss}{weights for different quantiles in calculating the cv error. Default is equal weight.}

\item{foldid}{indices of pre-split testing obervations}

\item{...}{other arguments for \code{gq.cv.pen}}
}
\value{
The full solution path is returned. It also returns the vector of CV score 
as well as the optimal lambda values in terms of min and 1se of the CV error. 
\item{beta}{The estimated coefficients for all lambdas, stored in sparse matrix format, where each column corresponds to a lambda.}
\item{lambda}{The sequence of lambdas.}
\item{cv_all}{An ntau*nlambda matrix of all values of CV error for each tau and each lambda.}
\item{err_all}{Errors for all \code{nfolds} folds evaluation. Has row of (ntau*nlambda) and column of nfolds.}
\item{lambda_min}{The optimal lambda that minimizes the CV error}
\item{lambda_1se}{The largest lambda such that CV error is within 1 standard error of the minimum CV error.}
\item{cv_min}{The value of CV error corresponding to \code{lambda_min}.}
\item{cv_1se}{The value of CV error corresponding to \code{lambda_1se}.}
\item{foldid}{The vector of indices for k folds split.}
\item{cvup}{CV error + 1 standard error}
\item{cvlo}{CV error - 1 standard error}
\item{n.nonzero.beta}{The number of selected covariates for each lambda.}
\item{eachtau}{A summary table of optimal lambdas with respect to each individual tau.}
}
\description{
Title Cross validation for consistent variable selection across multiple quantiles
}
\examples{
\dontrun{ 
n<- 200
p<- 20
X<- matrix(rnorm(n*p),n,p)
y<- -2+X[,1]+0.5*X[,2]-X[,3]-0.5*X[,7]+X[,8]-0.2*X[,9]+rt(n,2)
taus <- seq(0.1, 0.9, 0.2)
cvfit<- rq.gq.cv.pen(x=X, y=y, tau=taus)
}

}
